next block displays proper color to UI
I thought it was a good idea before to use the block type to set the color in
the UI. There were 7 blocks and 7 colors. It worked, and made it easier to
identify which block was next without looking too hard.

Whatever, "next/save" blocks in UI now present their proper (random) color to
the user. If it displays red, the next block will be red.

***

Don't display highscores if a user wants to save/quit.