Blocks are stored in a linked list.
So, I wanted more than just the 3 blocks we had previously(current, hold,
next). I decided a linked list would be the simplest way to hold more than 3
without drastically changing too much of the game.

As is:
We create 7 blocks in a linked list at the beginning of the game.

The first element in the list is *always* the HOLD piece.
The second element is the currently falling piece on the board.
The remaining five elements are the next blocks in order to appear on the
board.

Holding a piece simply swaps the first and second elements. And dropping a
piece moves it to the end of the list and randomizes it.

--

Other miscellaneous,
	comment updates,
	c89-ification(a bit),
	slight refactor(hopefully no logical changes.),
	made drawing more efficient, we no longer redraw the static text and
some parts of the board.

--

Actually I think the screen/drawing changes are more significant than the
logical block list changes.

-- TODO

A bit more still needs to be done in this branch, off the top of my head:
	write the code to draw the next pieces to the screen.
	I'm sure some other stuff.