Documentation is todo! I swear.
It's too much of a pain to constantly be keeping it up to date so early in the
game. So I slacked off.

Anyway, by 1.0 I promise it will be up-to-date and accurate and beautiful.

Unit tests are still something that I'm struggling with. I don't know exactly
what needs to be tested. And I think they're primarilly intended to reduce
introduced logical bugs during code-refactors. But, once I'm finished with this
and have moved on I'm not ever going to change the code. In any case, they
won't *find* bugs any better than e.g. valgrind to gdb.

Don't know :/