Changed algorithm in src/bag.c to be more efficient.
Relatively small change to the algorithm in src/bag.c.

The problem is to put the blocks into the bag randomly.

Before, we picked a random number and checked if it was available and if it was
we added it to the bag, then marked it dirty. Then we guessed again, etc. This
was very inefficient and wasted lots of cycles.

Now, we pick a random number up to the number of pieces remaining, and use that
number to select that element in the list of available pieces.

Example: If the bag contained 3 pieces: T, O, S, we would choose a number
between [1, 3] and then search through the list of remaining blocks to find the
third element.

Also I drew an ASCII diagram. I think it looks atrocious.