Draw Next/Hold blocks verticle and in their own box.
This commit is ugly; I should have commited somewhere inbetween these two
logical changes.

The first moves the Next/Hold block lists to a verticle position and draws a
box around it, maybe some other stuff.

The second actually changes how we draw. I was getting an extremely annoying
flash as the screen would update. So now we don't redraw the block lists unless
they've changed. We also try not to use wrefresh() inplace to
wnoutrefresh()/doupdate(). Which should help the throughput. Finally, we don't
draw anything if the game is paused.