Switched to an event_based architecture. This is version 0.1.
Rationale:
I ran into this problem with lock delays, which comes into play whenever the
block hits the bottom. The lock delay is supposed to be a waiting period where
the user is allowed a single move for a short time(<1 second) before the game
adds the block to the game and moves on.

The problem essentially boiled down to being unable to tell a function to stop
executing code. Basically, the functions were too big. I had two threads and
two infinite loops happening concurrently. But that meant complexity, because
now I have to control access to memory.

Networking was a big motivator as well. In the current thread-based model, I
would have to create a new thread to listen for connections/communication with
the server, and then lock memory to apply any changes the server sends us.

It's much easier to have a single infinite loop(event loop), that waits for
events and tells other sections of code to do what it wants.

Effectively, this commit breaks up the two infinite loops into atomic ticks.
One controlling gravity, the other controlling user input. In the future a new
event will be added to handle server messages.

To gravity update needs to happen periodically, so I used a POSIX timer to
handle that. In this very early stage of development, the timer doesn't depend
on the level.

Future:
	Update timer depending on user level.
	Use select/pselect in main loop to wait for events.
	Continue work on game win/lose conditions.
	Work on scheduling/main loop code.