getnextline() changes, moved replace_home() here.
getnextline changes a bit internally, if the user passes invalid variables to
the function it will reset the file offset. Call
	getnextline(NULL, 0, NULL);
before using it with valid data.

getnextline returns a newly allocated buffer to the user, instead of just a
buffer offset.

It interprets a '\n' and '\r' characters as new-line characters.

It will replace new line characters with '\0'.